use dainst fork of express-pouchdb
This is necessary in order to be able to use the fix that allows attachments
to be served correctly when express-pouchdb runs in electron.

The installation is configured as a postinstall script since pouchdb-server
uses a monorepo and npm does not provide a way to directly install subfolders
of git repositories.

This configuration should be reverted and we should use the official
express-pouchdb again when the pull-request is accepted.

See: https://github.com/pouchdb/pouchdb-server/pull/263
Resolves: #6301