create projects non-destructively
before it was like that, that on creation of a db, a db.destroy got issued, mainly motivated to achieve a cleanup on e2e testing, which of course should happen by the test code itself. but also in production it seemed kind of odd that when you create a db that already exists, it can start out with data in it. realistically that should never happen, unless someone touches the config.json in the users app folder so that the project is not listed but phyisically there. but then, if we can do it non-destructively, like with the code of this commit, we prefer to do it like this, since it reduces the risk of data loss, however unlikely it is.