load originals imgs with sample data loader
during a bugfix where one could not reload at the images route due to unsufficient promise handling there was this line with unclear purpose. we decided to remove it. only today it became apparent then that this was for the loading of the original images. to make this clear i added comments this time.