make pencil part of fields tab of document view
we already found out that it can mislead when we are in images tab and have the impression that we can edit images with it. which of course could be done with jumping to the images tab, but we want to get rid of that one in favor of another solution anywa. the actual problem apart from that i tried to avoid here is that when the window gets to small, one of the tabs symbols gets pushed to a new line. i made set the window min size to 996 today and i want to avoid to make that bigger. removing the pencil and placing it into the fields tab helps us staying with 996.