add pencils to relations and images tabs
after inserting the pencil into the fields tab, it had to go to the relations tab as well. and then to the images tab.
from the relations tab one jumps to the relations tab of docedit, and from images to the images tab of docedit. that
feels so smooth and consistent with the behaviour of the rest of document view / docedit
that i think we can drop the idea of adding images directly via image tab of documentview instead of docedit.