make depicts invisible but instead jump to images via thumbs in docview
this makes a lot more sense and feels more smooth, even more so since when you jump from imageview via a relation link, you land on the images tab of document view anyway, from which you again can jump to the image again. up until now this circle worked only for the resources, but didn't for the images.