extract imagegridcomponent; extract image-widgets module
we have the image-grid builder that generates the rows. we have the image-grid-component-base, through
which 4 different components use it and share code. the new image-grid-component now was created so that
the display of the grid is always same-ish. the old image-grid-component is now image-overview-component.
the other thing is that a new module was created for all image related widgets since these are many. they
also do not belong to the images package because they are used elsewhere, whereas the image module refers only
to the image management section.