select corresponding docview tab after save in docedit
since we can already retain the corresponding tab when switching from docview to docedit, this is a logical addition. we retain the tab only if we go out by saving, though, otherwise it could be a bit irritating if one just has a look and goes out by clicking the X, expecting nothing to be changed. so nothing should be changed also outside of docedit