extract imageview module
this leaves us with smaller, more focused modules. investigating an issue where we want to jump from documentview to an image and then exit via X where we want to jump back to resources with documentview revealed that imageview is in fact independent of imageoverview. this also leaves us with a nice dependency graph where images depends on imageview, imageview depends on docedit, and images doesn't anymore.