show project configuration errs in console
before the change to the direct injection of project configuration we showed the errors via messages. this does not work anymore because we stop the setup process of the app immediately. printing the errs on console should be totally fine though, since we deliver the app always with a conf.