Remove suspiciously problematic code.
I wanted to fix the todo and review the commment. I saw then, that the logic worked as follows, that if one liesWithin target is missing, it deletes all the liesWithin of that document silently. Silently due to the fact that the return value in form of msgWithParams, most probably a refactoring artifact and once shown to the user, was not send to messages. I cannot image the cause why we treated this relation differently at that point in the program. Data incosinstecies may happen, i.e. when synchronizing, yes, but why is it more relevant to the liesWithin relation than to any other? Since this is a bit rough behaviour, the reason is not documented, and on top of that it is not tested, I decided to take that code out.