Ignore empty array objects and append new entries.
This is for support of multiple import runs from the same import files. Throws an error if The resulting array would contain empty entries. As for example when the field from the import would contain array objects at indices 2 and 3, and the original array would contain an object only at the index 0. There would be no object at index 1. We want to make sure the indices match at a reimport, which is why we do not delete entries or shuffle entries around in other ways.