Do not collapse null fields in deletion mode.
This is to avoid situations where users specify null fields with the intention to delete, but then the whole object gets interpreted as to be deleted. Deletion should only work on a per field basis. So it could be that the original record's object still contains fields which are not specified by the null fields. These should be retained and only the specified ones should get deleted.