Merge ProjectCategories into ProjectConfiguration
With Labels and other concerns properly factored out, ProjectConfiguraton was concerned with providing relations related services, and also providing the forest, which got mainly consumed by ProjectCategories. The calls however were a bit ugly. And the relation related services were kind of the same nature as the things in ProjectCategories. Namely knowledge about the constraints and meanings of the relations and categories in relation to the application behaviour. So it makes for a good unified and cohesive service now.