Move functions from Relations to Resource
This makes things more cohesive. For one, resources have fields and relations (which we increasingly tend to treat more and more the same, in context of configuration), but we do not (but could) have Fields, only Relations. Now Resource has for example getDifferingFields and getDifferingRelations. The cohesiveness could be seen by being able to make a couple of functions private.