Sketch, Filter works basically
The sketching of a drawing has been integrated as an alternative search for similars. Works only with resnet though and NOT good!

Literature0 (the first and main of all literatureferences in the literature list) has been forwarded from backend to frontend. Can be filtered now.

Approaching now display of publication overview, using elasticsearch "agg"