Remove `/api` and add `/ui` route
The JSON API is now the default route, keeping the newer `/projects` route aligned with the `/files` and `/db` routes used for syncing.

Modules for the API still reside in `/api` directories. Some Phoenix internal plumbing has to be updated before we could move those in the main directory and the HTML related controllers/views/template into a `ui` subdirectory.