Fix problem caused by missing url decoding
When reading filters out from the url, the query params need to be decoded. They will serve as keys in ExistingFilters's dropdownMap and when they don't match up with the keys there, the app crashes /typical null pointer while trying to read further properties from the now undefined values/. A nuance with url param decoding: the + sign in the query params part /not in the url itself/ means 'space', while 'space' itself would be coded as '%2b'.