Keep original resolutions for map layer images in seed project
The images in the seed project were initially scaled down in order to not increase the repository size too much. Because of the change to the libvips preprocessing pipeline, the scaled down images now introduce errors because the geometry and other metadata in the database does not match the actual file width/height.