Fix json "icon"
There is no icon for curly braces, we just use curly braces as text. With the Phoenix Live view update the old implementation will get interpreted as an empty string interpolation, thus not showing any characters in the final html.

Instead we now pass a "{}" to be interpolated, which looks kinda hacky but is probably still more readable than adding the escaped representation of '{' and '}' ?