Split map feature highlight and tooltip creation
This also moves the html template used for the tooltip into the hook and extends the hook in way that we might show the names of all documents that are found at that coordinate. For the time being the map only shows one document and a click on the map will then open that document. In the future we might show a list and a click starts a search matching that coordinate?